Add prerelease workflow (#293)
We currently have a rather unreliable release cycle and sometimes wait
for several features to be completed before we bundle them into one
release. This can be cumbersome to users who wait for a particular
feature to be released.
This PR therefore introduces a workflow ~~that automatically pushes a
new `-beta` version to npmjs whenever changes land in `main`.
Prereleases can be skipped by adding `[skip prerelease]` to the merge
commit head, to avoid flooding users with new betas for every tiny
change.~~ to conveniently release a `beta` version to npmjs when
triggered manually.

Caveat: if we decide to employ this workflow we would have to always
bump the package version to the next expected number right _after_ a
release. We currently only bump the version number right _before_ a
release. But since the version number for the prereleases are derived
from the current version found in package.json, that would create a beta
version for the last released version, which doesn't make sense.