Add hashing (#513)
Setting `cds.typer.cache = 'blake2s256'` hashes the extended model and
compares it to the hash of the previous run read from disk (if it
exists). If the model has not changed, cds-typer will end the run
immediately at this point.

cds-typer currently needs the inferred and xtended flavour of a model
for a complete run. So when actually running, it performs `cds.load`
twice, considerably increasing runtime.
With the `cache` setting active, it loads the xtended model first,
checks the hash, and if the hash has not changed, does not perform the
second `cds.load`. So we should see a runtime improvement of about 50%
for the initial model loading when the cache is hit.

File IO is also cut down in case of a cache hit, as no files have to be
written.

A quick test with a model of ~8MB in size showed that the time went from
~13s on the initial run when no `@cds-models` existed, down to ~5s when
cds-typer would stop immediately after finding there was nothing to do.

```
➜ time cds-typer "*"
cds-typer "*"  13.03s user 1.12s system 133% cpu 10.613 total
➜ time cds-typer "*"
cds-typer "*"  5.35s user 0.32s system 136% cpu 4.159 total
```

The currently used algorithm
[blake2](https://en.wikipedia.org/wiki/BLAKE_(hash_function)#BLAKE2) is,
to my knowledge, the fastest hash algorithm available in the `crypto`
module. I deliberately left the config option open to using a different
algorithm in the future.

---------

Co-authored-by: Christian Georgi <chgeo@users.noreply.github.com>